ルートシグネチャの改良とその他
・パイプラインステート同様、ルートシグネチャをマネージャーで管理するようにし、一度だけ生成して保存し、使いまわせるようにした。
・いくつかのファイルで警告が出ていたので解決した。

+++ Job_Hunting_3DGame/Job_Hunting_3DGame/TitleScene.cpp
@@ -301,8 +301,8 @@ void TitleScene::ImGui_WaterMesh()
 
 		if (waterMesh)
 		{
-			static int gridSize = waterMesh->GetGridSize();
-			ImGui::DragInt("GridSize", &gridSize, 2.0f, 64, 1024);
+			static float gridSize = waterMesh->GetGridSize();
+			ImGui::DragFloat("GridSize", &gridSize, 2.0f, 64.0f, 1024.0f);
 			if (ImGui::Button("Apply Grid Size"))
 			{
 				waterMesh->Update_GridSize(gridSize);
